+++ NEWS.md
@@ -3,6 +3,7 @@
 *   Added the colCumSum function for fast computation of cumulative sums for each column in a matrix.
 *   Fixed a bug in auc 
 *   Fixed a bug in power_mcnemar_test and modified the default arguments [Thanks to John Lockwood]
+*   Added residual_plot to generate a residual plot using ggplot2 [Thanks to Søren Højsgaard]
 
 # MESS 0.5.5